requirements.txt
fastapi
uvicorn
langchain_groq
sentence-transformers
faiss-cpu
groq